requirements.txt
ddt==1.3.1
PyYAML==5.3.1
xlsxwriter==1.2.8
colorama==0.4.3
requests==2.23.0